Agrego dtype

+++ evaluate.py
@@ -26,7 +26,7 @@ def main():
     results = museval.EvalStore()
     with torch.no_grad():
         for track in tqdm.tqdm(mus):
-            signal = torch.as_tensor(track.audio.T).to(device)
+            signal = torch.as_tensor(track.audio.T, dtype=torch.float32).to(device)
             _, estimate = network(signal.unsqueeze(0))
             estimate = estimate.cpu()[0, ...]
             score = museval.eval_mus_track(track.targets[args.target].audio.T, estimate)
